Move the NetworkInteraction execution helpers.
This commit moves the old `network-itneractions.ts` module to the
FutureProcessor's helper module, as it's only intended to be used there.
